requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
requests==2.31.0
openai==1.30.1
gTTS==2.5.1
pygame==2.5.2
python-dotenv==1.0.1
